==> Aspect Aware Embeddings/01_Compute_Embeddings_bge-large-en-v1.5.ipynb <==
!pip install sentence-transformers

import pandas as pd
from google.colab import drive
import os
from sentence_transformers import SentenceTransformer
import pickle

drive.mount('/content/gdrive')
csv_path = '/content/gdrive/My Drive/Colab Data'

race_df_train = pd.read_csv(os.path.join(csv_path, 'race_train.csv'))
race_df_train

race_df_validation = pd.read_csv(os.path.join(csv_path, 'race_validation.csv'))
race_df_validation

race_df_test = pd.read_csv(os.path.join(csv_path, 'race_test.csv'))
race_df_test

dolly_cqa_df = pd.read_csv(os.path.join(csv_path, 'dolly_cqa.csv'))
dolly_cqa_df

model = SentenceTransformer('BAAI/bge-large-en-v1.5')

embeddings_train = model.encode(race_df_train['text'].to_list(), batch_size=128, show_progress_bar=True, normalize_embeddings=False)
print(embeddings_train.shape)

with open(os.path.join(csv_path, 'embeddings_train_bge_large.pickle'), 'wb') as f:
  pickle.dump(embeddings_train, f)

embeddings_validation = model.encode(race_df_validation['text'].to_list(), batch_size=128, show_progress_bar=True, normalize_embeddings=False)
print(embeddings_validation.shape)

with open(os.path.join(csv_path, 'embeddings_validation_bge_large.pickle'), 'wb') as f:
  pickle.dump(embeddings_validation, f)

embeddings_test = model.encode(race_df_test['text'].to_list(), batch_size=128, show_progress_bar=True, normalize_embeddings=False)
print(embeddings_test.shape)

with open(os.path.join(csv_path, 'embeddings_test_bge_large.pickle'), 'wb') as f:
  pickle.dump(embeddings_test, f)

embeddings_dolly = model.encode(dolly_cqa_df['text'].to_list(), batch_size=128, show_progress_bar=True, normalize_embeddings=False)
print(embeddings_dolly.shape)

with open(os.path.join(csv_path, 'embeddings_dolly_bge_large.pickle'), 'wb') as f:
  pickle.dump(embeddings_dolly, f)
